Implement SimpleHammer
Created SimpleHammer.java file that extends BasicWeapon and interfaces
with Weapon.

SimpleHammer initializes the DAMAGE to 35 hitpoints and overrides
hit() such that if the armor is less than 30, the hammer will ignore
armor points, otherwise armor reduces impact of DAMAGE.

WeaponFactory.java is updated to include SimpleHammer in the switch.

RoleGame.java is updated to test the functionality of the new weapon
class.